requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
absl-py=0.15.0=pyhd3eb1b0_0
aiohttp=3.8.1=py39hca72f7f_1
aiosignal=1.2.0=pyhd3eb1b0_0
appdirs=1.4.4=pypi_0
async-timeout=4.0.1=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
audioread=2.1.9=pypi_0
black=22.3.0=pypi_0
blas=1.0=mkl
blinker=1.4=py39hecd8cb5_0
brotlipy=0.7.0=py39h9ed2024_1003
bzip2=1.0.8=h1de35cc_0
c-ares=1.18.1=hca72f7f_0
ca-certificates=2022.4.26=hecd8cb5_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2021.10.8=py39hecd8cb5_2
cffi=1.15.0=py39hc55c11b_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py39hecd8cb5_0
cryptography=3.4.8=py39h2fd3fbb_0
cycler=0.11.0=pypi_0
dataclasses=0.8=pyh6d0b6a4_7
decorator=5.1.1=pypi_0
ffmpeg=4.2.2=h97e5cf8_0
ffmpeg-python=0.2.0=pypi_0
fonttools=4.33.3=pypi_0
freetype=2.11.0=hd8bbffd_0
frozenlist=1.2.0=py39hca72f7f_0
future=0.18.2=py39hecd8cb5_1
gettext=0.21.0=h7535e17_0
giflib=5.2.1=haf1e3a3_0
gmp=6.2.1=h23ab428_2
gnutls=3.6.15=hed9c0bf_0
google-auth=2.6.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
grpcio=1.42.0=py39ha29bfda_0
icu=58.2=h0a44026_3
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.3=py39hecd8cb5_0
intel-openmp=2021.4.0=hecd8cb5_3538
joblib=1.1.0=pypi_0
jpeg=9e=hca72f7f_0
jsonschema=4.5.1=pypi_0
kiwisolver=1.4.2=pypi_0
lame=3.100=h1de35cc_0
lcms2=2.12=hf1fd2bf_0
libcxx=12.0.0=h2f01273_0
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libiconv=1.16=hca72f7f_2
libidn2=2.3.2=h9ed2024_0
libopus=1.3.1=h1de35cc_0
libpng=1.6.37=ha441bb4_0
libprotobuf=3.20.1=h8346a28_0
librosa=0.9.1=pypi_0
libtasn1=4.16.0=h9ed2024_0
libtiff=4.2.0=h87d7836_0
libunistring=0.9.10=h9ed2024_0
libuv=1.40.0=haf1e3a3_0
libvpx=1.7.0=h378b8a2_0
libwebp=1.2.2=h56c3ce4_0
libwebp-base=1.2.2=hca72f7f_0
libxml2=2.9.12=hbf8cd5e_1
llvm-openmp=12.0.0=h0dcd299_1
llvmlite=0.38.0=pypi_0
lz4-c=1.9.3=h23ab428_1
markdown=3.3.4=py39hecd8cb5_0
matplotlib=3.5.2=pypi_0
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py39h9ed2024_0
mkl_fft=1.3.1=py39h4ab4a9b_0
mkl_random=1.2.2=py39hb2f4e1b_0
multidict=5.2.0=py39hca72f7f_2
musdb=0.4.0=pypi_0
museval=0.4.0=pypi_0
mypy-extensions=0.4.3=pypi_0
ncurses=6.3=hca72f7f_2
nettle=3.7.3=h230ac6f_1
numba=0.55.1=pypi_0
numpy=1.21.5=py39h2e5f0a9_2
numpy-base=1.21.5=py39h3b1a694_2
oauthlib=3.2.0=pyhd3eb1b0_0
openh264=2.1.1=h8346a28_0
openssl=1.1.1o=hca72f7f_0
packaging=21.3=pypi_0
pandas=1.4.2=pypi_0
pathspec=0.9.0=pypi_0
pesq=0.0.4=pypi_0
pillow=9.0.1=py39hde71d04_0
pip=21.2.4=py39hecd8cb5_0
platformdirs=2.5.2=pypi_0
pooch=1.6.0=pypi_0
protobuf=3.20.1=py39he9d5cce_0
pyaml=21.10.1=pypi_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pydeprecate=0.3.2=pypi_0
pyjwt=2.1.0=py39hecd8cb5_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.9=pypi_0
pyrsistent=0.18.1=pypi_0
pysocks=1.7.1=py39hecd8cb5_0
pystoi=0.3.3=pypi_0
python=3.9.12=hdfd78df_0
python-dateutil=2.8.2=pypi_0
pytorch=1.11.0=py3.9_0
pytorch-lightning=0.8.5=py_0
pytz=2022.1=pypi_0
pyyaml=6.0=py39hca72f7f_1
readline=8.1.2=hca72f7f_1
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
resampy=0.2.2=pypi_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-learn=1.1.0=pypi_0
scipy=1.7.3=py39h8c7af03_0
setuptools=61.2.0=py39hecd8cb5_0
simplejson=3.17.6=pypi_0
six=1.16.0=pyhd3eb1b0_1
soundfile=0.10.3.post1=pypi_0
sqlite=3.38.3=h707629a_0
stempeg=0.2.3=pypi_0
tensorboard=2.9.0=pyhd8ed1ab_0
tensorboard-data-server=0.6.0=py39h5896577_0
tensorboard-plugin-wit=1.6.0=py_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.11=h3fd3227_1
tomli=2.0.1=pypi_0
torch-tb-profiler=0.4.0=pypi_0
torchaudio=0.11.0=py39_cpu
torchmetrics=0.8.2=pypi_0
torchvision=0.12.0=py39_cpu
tqdm=4.64.0=py39hecd8cb5_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.9=py39hecd8cb5_0
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
x264=1!157.20191217=h1de35cc_0
xz=5.2.5=hca72f7f_1
yaml=0.2.5=haf1e3a3_0
yarl=1.6.3=py39h9ed2024_0
zipp=3.8.0=py39hecd8cb5_0
zlib=1.2.12=h4dc903c_2
zstd=1.4.9=h322a384_0
